% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{calculateGUI}
\alias{calculateGUI}
\title{A graphical user interface for the package rankFD}
\usage{
calculateGUI()
}
\description{
This function provides a graphical user interface for calculating rank-based 
statistical tests in general factorial designs.
}
\details{
The function produces a GUI for the calculation of the test statistics and 
for plotting. Data can be loaded via the "load data" button. The formula 
and the significance level alpha (default: 0.05) need to be specified.
One can choose between two different null hypotheses to be tested as well as
weighted or unweighted effects.
If the plot option is chosen, an additional window opens containing
information on the plots.
}

