% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Coal}
\alias{Coal}
\title{Coal Acidity}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{Acidity}{resulting acidity values}
  \item{NaOH}{the NaOH concentration}
  \item{Type}{three different types of coal: "Morwell", "Yallourn" and "Maddingley"}
}}
\source{
Hollander, M., Wolfe, D. A., Chicken, E. (2014) \emph{Nonparametric Statistical Methods}. Wiley Series in Probability and Statistics.

        Sternhell, S. (1958) Chemistry of brown coals VI: Further aspects of the chemistry of hydroxyl groups in Victorian brown coals. Australian Journal of Applied Science \bold{9}, 375--379.
}
\usage{
data(Coal)
}
\description{
Coal acidity values determined under each of three NaOH concentration levels for two different samples from each type of coal
}
\keyword{datasets}
