% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Muco}
\alias{Muco}
\title{Half-Time of Mucociliary Clearance}
\format{A data frame with 14 rows and 2 variables:
\describe{
  \item{HalfTime}{Half-Time of Mucociliary clearance, assessed from the rate of removal of dust}
  \item{Disease}{normal subjects, subjects with obstructive airways disease (OAD) and subjects with asbestosis}
}}
\source{
Hollander, M., Wolfe, D. A., Chicken, E. (2014) Nonparametric Statistical Methods. Wiley Series in Probability and Statistics.

        Thomson, M. L. and Short, M. D.(1969) Mucociliary function in health, chronic obstructive airway disease, and asbestosis. Journal of Applied Physiology \bold{26}, 535--539.
}
\usage{
data(Muco)
}
\description{
Mucociliary efficiency was assessed from the rate of removal of dust in three different groups of subjects
}
\keyword{datasets}
