% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{WMWSSP}
\alias{WMWSSP}
\title{Sample size calculation for the Wilcoxon-Mann-Whitney test.}
\usage{
WMWSSP(x1, x2, alpha, power, t)
}
\arguments{
\item{x1}{advance information for the first group}

\item{x2}{advance information for the second group}

\item{alpha}{two sided type I error rate}

\item{power}{power with the sample sizes of each group}

\item{t}{proportion of subjects in the first group}
}
\value{
Returns a data frame
}
\description{
Sample size calculation for the Wilcoxon-Mann-Whitney test.
}
\examples{
x1 = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 2)
x2 = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 3)
WMWSSP(x1,x2,0.05,0.8,0.5)
}
\references{
Brunner, E., Bathke A. C. and Konietschke, F. Rank- and Pseudo-Rank Procedures in Factorial Designs - Using R and SAS. Springer Verlag.
}
